very big commit: new (refcounted) base_node objects based on bgeot::small_vector<T>. New bgeot::kdtree class for point_in_box searches on a large set of points. Speed of bgeot::geotrans_inv was improved and the code was cleaned up. Interpolations should also be faster. Minor fixes.

+++ tests/geo_trans_inv.pl
@@ -27,7 +27,7 @@ while (<F>) {
     print $_, <F>;
   }
 }
+close(F);
+if ($?) { `rm -f $tmp`; exit(1); }
 if ($er == 1) { `rm -f $tmp`; exit(1); }
-`./geo_trans_inv $tmp`;
-`rm -f $tmp`; 
-if ($?) { exit(1); }
+`rm -f $tmp`;
